+++ test/BusinessPoliciesManagement/Services/ServiceTest.php
@@ -39,7 +39,7 @@ class ServiceTest extends \PHPUnit_Framework_TestCase
             'authToken' => '321',
             'credentials' => ['appId' => '', 'certId' => '', 'devId' => ''],
             'globalId' => '999',
-            'handler' => $h
+            'httpHandler' => $h
         ]);
 
         $s->testOperation();
@@ -68,7 +68,7 @@ class ServiceTest extends \PHPUnit_Framework_TestCase
             'authToken' => '321',
             'credentials' => ['appId' => '', 'certId' => '', 'devId' => ''],
             'globalId' => '999',
-            'handler' => $h
+            'httpHandler' => $h
         ]);
 
         $s->testOperation();