+++ test/Types/AttachmentTest.php
@@ -18,9 +18,9 @@ class AttachmentTest extends \PHPUnit_Framework_TestCase
          * The Handler contains properties that will be set when the service
          * makes the request. We can test these properties to check what the service is passing.
          */
-        $this->handler = new Handler();
+        $this->httpHandler = new Handler();
         // BaseService is abstract so use class that is derived from it for testing.
-        $this->service = new Service(['handler' => $this->handler]);
+        $this->service = new Service(['httpHandler' => $this->httpHandler]);
         $this->request = new ComplexClass();
         $this->request->integer = 123;
         $this->request->string = 'a string';
@@ -54,24 +54,24 @@ class AttachmentTest extends \PHPUnit_Framework_TestCase
     {
         $this->request->attachment('ABC123', 'image/jpeg');
         $this->service->foo($this->request);
-        $this->assertArrayHasKey('fooHdr', $this->handler->headers);
-        $this->assertEquals('foo', $this->handler->headers['fooHdr']);
-        $this->assertArrayHasKey('Content-Type', $this->handler->headers);
-        $this->assertEquals('multipart/related;boundary=MIME_boundary;type="application/xop+xml";start="<request.xml@devbay.net>";start-info="text/xml"', $this->handler->headers['Content-Type']);
-        $this->assertArrayHasKey('Content-Length', $this->handler->headers);
-        $this->assertEquals(strlen($this->requestXml), $this->handler->headers['Content-Length']);
+        $this->assertArrayHasKey('fooHdr', $this->httpHandler->headers);
+        $this->assertEquals('foo', $this->httpHandler->headers['fooHdr']);
+        $this->assertArrayHasKey('Content-Type', $this->httpHandler->headers);
+        $this->assertEquals('multipart/related;boundary=MIME_boundary;type="application/xop+xml";start="<request.xml@devbay.net>";start-info="text/xml"', $this->httpHandler->headers['Content-Type']);
+        $this->assertArrayHasKey('Content-Length', $this->httpHandler->headers);
+        $this->assertEquals(strlen($this->requestXml), $this->httpHandler->headers['Content-Length']);
     }
 
     public function testXmlIsCreated()
     {
         $this->request->attachment('ABC123', 'image/jpeg');
         $this->service->foo($this->request);
-        $this->assertEquals($this->requestXml, $this->handler->body);
+        $this->assertEquals($this->requestXml, $this->httpHandler->body);
     }
 
     public function testResponseIsReturned()
     {
-        $this->handler->returnAttachment = true;
+        $this->httpHandler->returnAttachment = true;
         $response = $this->service->bar($this->request);
         $this->assertInstanceOf('\DTS\eBaySDK\Test\Mocks\ComplexClass', $response);
         $this->assertEquals(123, $response->integer);